Authentication Updates:
- Made a LocalAuthGuard instead of AuthGuard('local')
- Created an AuthenticatedGuard for authenticated users
- Created a SessionSerializer
- main.ts cleanup
- Updated and fixed all routes / apis
- Now, log in works to add a session cookie, log out removes the cookie, when logged in, we can access AuthenticatedGuard routes! So excited!!

+++ src/definitions/services/definitions/definitions.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Definition } from '../../../typeorm/entities/definition';
 import { Repository, DeleteResult, UpdateResult } from 'typeorm';
 import { CreateDefinitionDto } from '../../dtos/create-definition.dto';
 import { UpdateDefinitionDto } from '../../dtos/update-definition.dto';
+import { User } from '../../../typeorm/entities/user';
 
 @Injectable()
 export class DefinitionsService {
@@ -68,21 +69,44 @@ export class DefinitionsService {
   }
 
   async createDefinition(
+    user: User,
     createDefinitionDto: CreateDefinitionDto,
   ): Promise<Definition> {
     const newDefinition =
       this.definitionsRepository.create(createDefinitionDto);
+    newDefinition.user = user;
     return this.definitionsRepository.save(newDefinition);
   }
 
   async updateDefinitionById(
+    user: User,
     id: number,
     updateDefinitionDto: UpdateDefinitionDto,
   ): Promise<UpdateResult> {
+    const definition = await this.getDefinitionById(id);
+
+    if (!definition) {
+      throw new HttpException('Definition not found', HttpStatus.NOT_FOUND);
+    }
+    if (definition.userId !== user.id) {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.definitionsRepository.update(id, updateDefinitionDto);
   }
 
-  async deleteDefinitionById(id: number): Promise<DeleteResult> {
+  async deleteDefinitionById(user: User, id: number): Promise<DeleteResult> {
+    const definition = await this.definitionsRepository.findOne({
+      where: { id },
+    });
+
+    if (!definition) {
+      throw new HttpException('Definition not found', HttpStatus.NOT_FOUND);
+    }
+    if (definition.userId !== user.id) {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.definitionsRepository.delete(id);
   }
 }
